se añade añade elimina archivo al actualizar

+++ src/Exceptions/deleteFileException.ts
@@ -14,13 +14,9 @@ export class DeleteFileException implements ExceptionFilter {
 		const response = ctx.getResponse<Response>();
 		const request = ctx.getRequest<Request>();
 		const status = exception.getStatus();
-		console.log({
-			status,
-			exception: exception.message,
-			res: exception.getResponse(),
-			req: request.file,
-		});
-		fs.unlinkSync(request.file.path);
+		if(request.file){
+            fs.unlinkSync(request.file.path);
+        }
 		response.status(status).json({
 			status,
 			exception: exception.message,
